requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
async-timeout==4.0.2
attrs==21.4.0
bidict==0.22.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cx-Freeze==6.10
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.4
multidict==6.0.2
patchelf==0.14.5.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.27.1
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
